requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-prometheus==0.42b0
prometheus-client==0.19.0
groq==0.11.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
